+++ syncer-transmission/src/main/java/syncer/syncerservice/util/JDRedisClient/JDRedisJedisClient.java
@@ -119,20 +119,22 @@ public class JDRedisJedisClient implements JDRedisClient {
 
 
     @Override
-    public String restore(Long dbNum, byte[] key, int ttl, byte[] serializedValue) {
+    public String restore(Long dbNum, byte[] key, long ttl, byte[] serializedValue) {
         return jdJedisClientPool.restore(key,ttl,serializedValue);
     }
 
     @Override
-    public String restoreReplace(Long dbNum, byte[] key, int ttl, byte[] serializedValue) {
+    public String restoreReplace(Long dbNum, byte[] key, long ttl, byte[] serializedValue) {
         return jdJedisClientPool.restoreReplace(key,ttl,serializedValue);
 
     }
 
     @Override
-    public String restoreReplace(Long dbNum, byte[] key, int ttl, byte[] serializedValue, boolean highVersion) {
+    public String restoreReplace(Long dbNum, byte[] key, long ttl, byte[] serializedValue, boolean highVersion) {
         return jdJedisClientPool.restoreReplace(key,ttl,serializedValue,highVersion);
     }
+
+
 
     @Override
     public Object send(byte[] cmd, byte[]... args) {
