+++ syncer-transmission/src/main/java/syncer/syncerservice/util/JDRedisClient/JDRedisRedissonClient.java
@@ -110,17 +110,17 @@ public class JDRedisRedissonClient implements JDRedisClient {
 
 
     @Override
-    public String restore(Long dbNum, byte[] key, int ttl, byte[] serializedValue) {
+    public String restore(Long dbNum, byte[] key, long ttl, byte[] serializedValue) {
         return null;
     }
 
     @Override
-    public String restoreReplace(Long dbNum, byte[] key, int ttl, byte[] serializedValue) {
+    public String restoreReplace(Long dbNum, byte[] key, long ttl, byte[] serializedValue) {
         return null;
     }
 
     @Override
-    public String restoreReplace(Long dbNum, byte[] key, int ttl, byte[] serializedValue, boolean highVersion) {
+    public String restoreReplace(Long dbNum, byte[] key, long ttl, byte[] serializedValue, boolean highVersion) {
         return null;
     }
